requirements.txt
absl-py==2.1.0
accelerate==0.29.3
aiofiles==23.2.1
altair==5.3.0
annotated-types==0.6.0
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==23.2.0
audioread==3.0.1
babel==2.17.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
blis==0.7.11
cachetools==5.3.3
catalogue==2.0.10
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
cmake==3.28.3
comm==0.2.1
confection==0.1.4
contourpy==1.1.1
cycler==0.12.1
cymem==2.0.8
Cython==3.0.9
debugpy==1.8.1
decorator==5.1.1
deepspeed==0.17.1
defusedxml==0.7.1
diffusers==0.21.2
dm-tree==0.1.8
dnspython==2.6.1
einops==0.7.0
email_validator==2.2.0
entrypoints==0.4
exceptiongroup==1.2.0
execnet==2.0.2
executing==2.0.1
expecttest==0.1.3
fastapi==0.111.0
fastapi-cli==0.0.4
fastjsonschema==2.19.1
ffmpy==0.3.2
filelock==3.13.1
flash-attn @ git+https://github.com/Dao-AILab/flash-attention.git@4976650f743ece43bb159a55372557c777a4489b
fonttools==4.49.0
fqdn==1.5.1
fsspec==2024.2.0
gast==0.5.4
google-auth==2.28.2
google-auth-oauthlib==0.4.6
gradio==3.50.2
gradio_client==0.6.1
grpcio==1.62.1
h11==0.14.0
hf-xet==1.1.4
hjson==3.1.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.4
hypothesis==5.35.1
idna==3.6
igraph==0.11.4
imageio==2.35.1
importlib_metadata==8.5.0
importlib_resources==6.4.0
iniconfig==2.0.0
intel-openmp==2021.4.0
ipykernel==6.29.3
ipython==8.12.3
ipython-genutils==0.2.0
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
json5==0.9.22
jsonpointer==3.0.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter-tensorboard==0.2.0
jupyter_client==7.4.9
jupyter_core==5.7.1
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==2.3.2
jupyterlab-server==1.2.0
jupyterlab_pygments==0.3.0
jupytext==1.16.1
kiwisolver==1.4.5
langcodes==3.3.0
lark==1.1.9
lazy_loader==0.4
librosa==0.10.1
lightning-utilities==0.10.1
llvmlite==0.41.1
loguru==0.7.2
looseversion==1.3.0
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.6
mdit-py-plugins==0.4.0
mdurl==0.1.2
mistune==3.0.2
mkl==2021.1.1
mkl-devel==2021.1.1
mkl-include==2021.1.1
mock==5.1.0
mpmath==1.3.0
msgpack==1.0.8
murmurhash==1.0.10
nbclassic==1.2.0
nbclient==0.9.0
nbconvert==7.16.2
nbformat==5.9.2
nest-asyncio==1.6.0
networkx==3.1
ninja==1.11.1.1
notebook==6.4.10
notebook_shim==0.2.4
numba==0.58.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-dali-cuda120==1.35.0
nvidia-ml-py==12.575.51
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
nvidia-pyindex==1.0.9
oauthlib==3.2.2
opencv-python==4.12.0.88
opt-einsum==3.3.0
optree==0.10.0
orjson==3.10.5
overrides==7.7.0
packaging==23.2
pandas==2.0.3
pandocfilters==1.5.1
parso==0.8.3
peft==0.10.0
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
pkgutil_resolve_name==1.3.10
platformdirs==4.2.0
pluggy==1.4.0
polygraphy==0.47.1
pooch==1.8.1
preshed==3.0.9
prettytable==3.10.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
protobuf==3.19.0
psutil==7.0.0
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==16.1.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pybind11==2.11.1
pybind11-global==2.11.1
pycocotools @ git+https://github.com/nvidia/cocoapi.git@d99cbf3823588ef09a2721655f46e509ebafb3d7#subdirectory=PythonAPI
pycparser==2.21
pydantic==2.6.3
pydantic_core==2.16.3
pydub==0.25.1
Pygments==2.17.2
pynvml==11.5.3
pyparsing==3.1.2
pytest==8.0.2
pytest-flakefinder==1.1.0
pytest-rerunfailures==13.0
pytest-shard==0.1.2
pytest-xdist==3.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-hostlist==1.23.0
python-json-logger==3.3.0
python-multipart==0.0.9
pytorch-fid==0.3.0
pytorch-quantization==2.1.2
pytz==2025.2
PyWavelets==1.4.1
PyYAML==6.0.1
pyzmq==25.1.2
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.0.0
rich-toolkit==0.14.7
rpds-py==0.18.0
rsa==4.9
safetensors==0.4.3
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
semantic-version==2.10.0
Send2Trash==1.8.2
sentencepiece==0.1.99
shellingham==1.5.4
six==1.16.0
smart-open==6.4.0
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.12.1
soupsieve==2.5
soxr==0.3.7
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sphinx_glpi_theme==0.6
srsly==2.4.8
stack-data==0.6.3
starlette==0.37.2
sympy==1.12
tabulate==0.9.0
tbb==2021.11.0
tensorboard==2.9.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
terminado==0.18.0
texttable==1.7.0
thinc==8.2.3
threadpoolctl==3.3.0
tifffile==2023.7.10
timm==0.9.5
tinycss2==1.2.1
tokenizers==0.15.2
toml==0.10.2
tomli==2.0.1
toolz==1.0.0
torch==2.4.1
torchvision==0.19.1
tornado==6.4
tqdm==4.66.2
traitlets==5.9.0
transformer-engine @ git+https://github.com/NVIDIA/TransformerEngine.git@0fbc76af3733ae997394eaf82b78ff9c0498fe91
transformers==4.39.1
triton==3.0.0
typer==0.9.4
types-dataclasses==0.6.6
types-python-dateutil==2.9.0.20241206
typing_extensions==4.10.0
tzdata==2024.1
ujson==5.10.0
uri-template==1.3.0
urllib3==2.2.3
uvicorn==0.30.1
uvloop==0.19.0
wasabi==1.1.2
watchfiles==0.22.0
wcwidth==0.2.13
weasel==0.3.4
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==11.0.3
Werkzeug==3.0.1
xdoctest==1.0.2
zipp==3.20.2
